review fixups
